test: isolate full embedded-cluster lifecycles
