+++ pkg/tests/issues/isolated/issue_25782_broadcast_budget_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package issues
+package isolated
 
 import (
 	"context"
