+++ pkg/tests/issues/isolated/joinspill_helpers_test.go
@@ -1,0 +1,25 @@
+// Copyright 2026 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package isolated
+
+import "github.com/matrixorigin/matrixone/pkg/tests/issues/internal/joinspill"
+
+var (
+	execJoinSpillSQL           = joinspill.ExecSQL
+	queryJoinSpillText         = joinspill.QueryText
+	patchJoinSpillStats        = joinspill.PatchStats
+	patchJoinSpillStatsWithNDV = joinspill.PatchStatsWithNDV
+	resetOptimizerHintsOnCN    = joinspill.ResetOptimizerHintsOnCN
+)
